feat(detection): add high-density multistream
