--- examples/object-detection/16-stream-object-detector/src/common/config-24x720p20.yaml
@@ -1,66 +1,0 @@
-# Proven profile: 24 x 720p @ 20 FPS (q4/i1/m4).
-model:                       # Detector model selection. Default run is pinned to YOLO26n.
-  path: yolo26n_raw_supported_einsum_mpk.tar.gz # Resolved relative to this config file.
-  labels: coco_label.txt      # Resolved relative to this config file.
-
-streams:                     # RTSP inputs, one entry per camera. Default profile: 24 x 720p @ 20 fps.
-  - rtsp://192.168.1.9:25689/src1
-  - rtsp://192.168.1.9:25689/src2
-  - rtsp://192.168.1.9:25689/src3
-  - rtsp://192.168.1.9:25689/src4
-  - rtsp://192.168.1.9:25689/src5
-  - rtsp://192.168.1.9:25689/src6
-  - rtsp://192.168.1.9:25689/src7
-  - rtsp://192.168.1.9:25689/src8
-  - rtsp://192.168.1.9:25689/src9
-  - rtsp://192.168.1.9:25689/src10
-  - rtsp://192.168.1.9:25689/src11
-  - rtsp://192.168.1.9:25689/src12
-  - rtsp://192.168.1.9:25689/src13
-  - rtsp://192.168.1.9:25689/src14
-  - rtsp://192.168.1.9:25689/src15
-  - rtsp://192.168.1.9:25689/src16
-  - rtsp://192.168.1.9:25689/src17
-  - rtsp://192.168.1.9:25689/src18
-  - rtsp://192.168.1.9:25689/src19
-  - rtsp://192.168.1.9:25689/src20
-  - rtsp://192.168.1.9:25689/src21
-  - rtsp://192.168.1.9:25689/src22
-  - rtsp://192.168.1.9:25689/src23
-  - rtsp://192.168.1.9:25689/src24
-
-input:                       # RTSP input transport settings.
-  tcp: true                  # Use TCP for RTSP transport.
-  latency_ms: 100            # RTSP receive latency in milliseconds.
-  drop_on_latency: true      # Keep live display current instead of preserving stale RTSP frames.
-  skip_rtsp_probe: true      # Use explicit caps below; avoids slow per-stream startup probes.
-  width: 1280                # Required when input.skip_rtsp_probe is true.
-  height: 720
-  fps: 20
-  decoder_buffers: 16        # Per-stream decoder output pool used by the validated profile.
-  decoder_input_buffers: 2   # Per-stream compressed-input pool.
-  decoder_tuning: auto
-
-runtime:                     # Runtime and profiling settings.
-  profile: false             # Leave hot-path profiling off for the live visualizer run.
-  warmup_frames: 30          # Drain this many frames before metadata/profile output.
-  initial_detection_timeout_ms: 90000 # First detection wait before treating startup as stalled.
-
-inference:                   # Detection controls.
-  workers: 1                 # One shared detector graph avoids dispatcher/resource contention.
-  fan_in_policy: every_frame # One pending frame/stream with bounded producer backpressure.
-  queue_depth: 4             # Terminal detection output and Run queue depth.
-  internal_queue_depth: 1    # One global model-stage handoff; not a per-camera decoded queue.
-  max_inflight_per_stream: 4 # Proven per-stream mux credit for approximately 20 detections/s.
-  min_score: 0.30            # Detection score threshold used by the validated Insight run.
-  nms_iou: 0.60              # NMS IOU threshold.
-  max_detections: 50         # Max detections to keep per frame after NMS.
-
-output:                      # Insight video and metadata output.
-  insight:                   # Live video and metadata destination.
-    host: 192.168.1.9        # Host running the Insight receiver/viewer.
-    video_port_base: 9000    # Base UDP port for video channels.
-    metadata_port_base: 9100 # Base UDP port for metadata channels.
-    max_visible_streams: 24  # Publish all 24 video and metadata channels to Insight.
-    sync_delay_ms: 400       # CPU-side H.264 delay; no additional EV74 frame queue is allocated.
-  video_enabled: true        # Disable to publish metadata only.
