--- examples/object-detection/16-stream-object-detector/tests/test-scope.yaml
@@ -1,3 +1,0 @@
-unit:
-  python: true
-  cpp: true
