+++ examples/object-detection/high-density-multi-stream-object-detector/src/common/coco_label.txt
