+++ examples/object-detection/high-density-multi-stream-object-detector/src/common/config-24x720p20fps.yaml
@@ -1,8 +1,10 @@
-model:                       # Detector model selection. Default run is pinned to YOLO26n.
-  path: yolo26n_raw_supported_einsum_mpk.tar.gz # Resolved relative to this config file.
+# Profile: 24 x 720p @ 20 FPS (q4/i1/m4).
+model:                       # Detector model selection.
+  path: yolo26n-det-bf16-mla_tess-b1.tar.gz # Resolved relative to this config file.
   labels: coco_label.txt      # Resolved relative to this config file.
+  decode_type: yolo26
 
-streams:                     # RTSP inputs, one entry per camera. Default profile: 24 x 720p @ 20 fps.
+streams:                     # RTSP inputs, one entry per camera.
   - rtsp://192.168.1.9:25689/src1
   - rtsp://192.168.1.9:25689/src2
   - rtsp://192.168.1.9:25689/src3
