+++ examples/object-detection/high-density-multi-stream-object-detector/src/python/requirements.txt
